+++ test/oban/config_test.exs
@@ -24,15 +24,6 @@ defmodule Oban.ConfigTest do
       assert_valid(circuit_backoff: 10)
     end
 
-    test ":dispatch_cooldown is validated as a positive integer" do
-      assert_invalid(dispatch_cooldown: -1)
-      assert_invalid(dispatch_cooldown: 0)
-      assert_invalid(dispatch_cooldown: "5")
-      assert_invalid(dispatch_cooldown: 1.0)
-
-      assert_valid(dispatch_cooldown: 500)
-    end
-
     test ":crontab is validated as a list of cron job expressions" do
       assert_invalid(crontab: ["* * * * *"])
       assert_invalid(crontab: [["* * * * *", Fake]])
@@ -94,13 +85,14 @@ defmodule Oban.ConfigTest do
       assert_valid(prefix: "private")
     end
 
-    test ":queues are validated as atom, integer pairs" do
+    test ":queues are validated as atom, integer pairs or atom, keyword pairs" do
       assert_invalid(queues: %{default: 25})
       assert_invalid(queues: [{"default", 25}])
       assert_invalid(queues: [default: 0])
       assert_invalid(queues: [default: 3.5])
 
       assert_valid(queues: [default: 1])
+      assert_valid(queues: [default: [poll_interval: 50, limit: 1]])
 
       assert %Config{queues: []} = conf(queues: false)
     end